requirements.txt
asgiref==3.3.1
certifi==2020.11.8
cffi==1.14.4
cryptography==3.2.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.7.0
django-heroku==0.3.1
django-phonenumber-field==5.0.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
whitenoise==5.2.0
wincertstore==0.2
